requirements.txt
requests
python-dotenv
python-dateutil
paho-mqtt
